+++ back/src/__tests__/utils.test.ts
@@ -1,4 +1,4 @@
-import { getUid, sameDayMidnight, daysBetween } from "../utils";
+import { getUid, sameDayMidnight, daysBetween, base32Encode } from "../utils";
 
 test("getUid returns a unique identifier of fixed length", () => {
   const uid = getUid(10);
@@ -17,4 +17,16 @@ test("daysBetween should calculate the number of days between two dates", () => 
   const date2 = new Date("2019-10-01T09:00:00");
   const days = daysBetween(date1, date2);
   expect(days).toEqual(3);
+});
+
+describe("base32Encode", () => {
+  it("should base32Encode a number", () => {
+    expect(base32Encode(0)).toEqual("0");
+    expect(base32Encode(1)).toEqual("1");
+    expect(base32Encode(9)).toEqual("9");
+    expect(base32Encode(10)).toEqual("A");
+    expect(base32Encode(31)).toEqual("Z");
+    expect(base32Encode(32)).toEqual("10");
+    expect(base32Encode(Math.pow(32, 9) - 1)).toEqual("ZZZZZZZZZ");
+  });
 });
