+++ back/src/forms/readableId.ts
@@ -1,5 +1,5 @@
 import { format } from "date-fns";
-
+import { base32Encode } from "../utils";
 export enum ReadableIdPrefix {
   BSD = "BSD", // Bordereau de suivi des déchets dangereux "générique" (CERFA n° 12571*01)
   DASRI = "DASRI", // Bordereau de suivi des déchets d'activités de soins à risque infectieux
@@ -39,34 +39,4 @@ export default function getReadableId(prefix = ReadableIdPrefix.BSD) {
  */
 export function getRandomInt(max: number) {
   return Math.floor(Math.random() * Math.floor(max));
-}
-
-/**
- * base32-crockford
- * ================
- * Implement the alternate base32 encoding as described
- * by Douglas Crockford at: http://www.crockford.com/wrmg/base32.html.
- * He designed the encoding to:
- *   * Be human and machine readable
- *   * Be compact
- *   * Be error resistant
- *   * Be pronounceable
- * It uses a symbol set of 10 digits and 22 letters, excluding I, L O and U.
- */
-export function base32Encode(n: number): string {
-  const alphabet = "0123456789ABCDEFGHJKMNPQRSTVWXYZ".split("");
-  function encode(n: number, encoded = "") {
-    if (n > 0) {
-      const r = n % 32;
-      const q = (n - r) / 32;
-      const symbol = alphabet[r];
-      return encode(q, symbol + encoded);
-    }
-    return encoded;
-  }
-  if (n === 0) {
-    return "0";
-  } else {
-    return encode(n);
-  }
 }
